serialize correctly; accounting for Infinities
This commit fixes up those instances in the core that were still
emplying `JSON.stringify` for serializing erps and values relating to
erps. This is faulty because `JSON.serialize` doesn't handle Infinities,
serializing both kinds as `null`.

I left one instance where it could be converted, but wasn't necessary --
in caching the function ids in IncrementalMH. It doesn't look like those
can be Infinities.
My simple tests don't seem to indicate that using a 'replacer' with
`JSON.stringify`, as `util.serialize` does affects performance, but I
left it out anyways.
If @dritchie is comfortable with it being changed, then we can make the
change.

Summary:

 - replace relevant instances of `JSON.stringify` with `util.serialize`
 - fix cache test so that it checks for correct handling of Infinities
 - minor corrections in rejection.js and pmcmc.js where `util.logHist`
   was being used without importing util.js

Related to: #162