Separate caching test into two tests.
The current test is hard to understand. It's also likely to generate
false positives as there are several unexpected values which e1 and e2
can take where the model returns true. (e.g. "isNaN({} + {})".) This
would not cause the test to fail as returning true with probability 1 is
within the histogram tolerance of most inference tests.

This new deterministic cache test would have failed prior to the recent
update which fixed serialized of ±Infinity.

This reverts the change made to the stochastic test made in 875812b2.