Fix Query handling and acceptance ratio in MCMC.
Prior to this commit, when using sequenceKernel, only the final kernel
in the sequence was taken into account when computing the acceptance
ratio.

Query handling had to be modified for a similar reason. i.e. It needs to
happen after each kernel in the sequence, not once at the end of the
sequence.