Source map only those errors that occur in eval.
Previously all errors (including those originating in plain JS) were
looked up in the source map. It was often the case that this happened to
work because the map doesn't contain entries for most line, column
number pairs. But when there was a match, it led to incorrect reported
of the origin of the error.

This commit removes the explicit check for errors occurring in native
code as those are a subset of the cases now caught by the 'eval' check.