Have indexOf follow JS behavior.
Returning a falsey value when the item is not found isn't super useful
since zero is falsey too. Rather than giving the impression that this
might be equivalent to something like _.contains when used as a
predicate, I think it's clearer to have this follow the behavior of the
JS version.