Remove indexOf header function.
Returning a falsey value when the element is not found is not useful as
a falsey value is also returned when the element is at index 0. Without
this modification this is just a wrapper around the JS indexOf. Since
this is also rarely used it was decided (#480) to remove it.