Improve the browser interface to run/compile.
The makes it possible to get our hands on the wrapped runner from the
browser. This is important because we need to manually resume
trampolining after doing work asynchronously.

It also gives us more flexibility in how we perform compilation in the
browser. e.g. I expect it to be easier to perform compilation in a web
worker after this change.